Response error propagated through exception flow.
This is necessary for control flow that needs to handle particular Error which may happen from diffrerent types of response.

Sadly, some responses use Error and the response structure itself to signal failure, so we actually need to propagate the reponse as well as the error in the exception path.
Hopfully, this will be fixed in a future protocol version.