Complete version of BatchClient.
Still pretty rough: cannot reuse the non-batch client response handlers due to the exception path being different.

Now that all the handlers had to be written to handle the response errors, we can reconsider using the ProtobufFilter to handle the response error.
Although that has the advantage of being much more consise.

It would be nice to write the BatchClient in terms of the normal Client, but the Future gets in the way. Alternatively, we could write the normal Client in terms
of the BatchClient, but that would make it heavier and would only be acceptable if it handles single-element batches as non-batch requests.