Change tx() signature to handle user priority
tx() providing a txKv to the transactional function which allowed the user to set a different user and priority on the tx. This caused the EndTransaction request to be sent with a potentially incorrect user and or priority.

This changes the signature of tx such that it knows about the user and priority of the tx and provides a new TxClient to the function.